requirements.txt
streamlit
chromadb
langchain
langchain-google-vertexai
pypdf
service